Update main.yml to fix out-of-date bundler
Attempts to build webpage after modifying nothing but `morea` files resulted in failed actions with this suggestion buried in the logs:

```
The last version of bundler (>= 0) to support your Ruby & RubyGems was 2.4.22. Try installing it with `gem install bundler -v 2.4.22`
```

An additional job that runs the `gem install bundler -v 2.4.22` command should - in theory - update the virtual machine used to compile the website and allow bundle to work properly.